+++ CHANGELOG.rst
@@ -30,6 +30,7 @@ Changed
 - Set max_page of 1000 for API and default page size to 25
 - Moved depend JS to static files
 - Change logging to file rotation for quickstart project
+- Improve Table component styling options
 
 Fixed
 ~~~~~